+++ server/src/routes/v1/users.routes.ts
@@ -1,10 +1,14 @@
 import { Router } from 'express';
-// import { listUsers, getUser, patchUser } from '../../controllers/user.controller';
-// import { validate } from '../../middleware/validate';
-// import { IdParamSchema } from '../../validators/common.schema';
+import { UserController } from '../../controllers/user.controller';
+import { validate } from '../../middleware/validate';
+import { GetUserSchema } from '../../validators/user.schema';
+import { auth } from '../../middleware/auth';
 
-export const router = Router();
+export const createUserRoutes = (controller: UserController) => {
+    const router = Router();
 
-// router.get('/', listUsers);
-// router.get('/:id', validate(IdParamSchema), getUser);
-// router.patch('/:id', validate(IdParamSchema), patchUser);
+    router.get('/me', auth, (req, res, next) => controller.getMe(req, res, next));
+    router.get('/:username', validate(GetUserSchema), (req, res, next) => controller.getUser(req, res, next));
+
+    return router;
+}